feat: :sparkles: TwoPointers: Is palindrome
Function to check if a string is a palindrome. For it we have a function with two pointers (start and end). It will iterate the string verifying if the last char equals to the first (and ++start and --end every loop). If in any moment it is different, we return false. Case not, we return true. Obs: the middlechar is when start and end are the same.