feat: :sparkles: Reverse String
Creating a algorithm to reverse strings. We are using the two pointers technique to iterate string and define the head (index 0) with the tail (index -1) value and vice versa (also put tail with head value). So this way, each new loop we revert the chars by defining index positions into the new array of runes